requirements.txt
flask==3.1.0
tensorflow==2.19.0
keras==3.9.2
numpy==2.1.3
Pillow==11.2.1
mlflow